+++ libs/firstmall-db/src/lib/fm-orders/fm-orders.service.ts
@@ -29,15 +29,15 @@ import {
 } from '@project-lc/shared-types';
 import dayjs from 'dayjs';
 import { SellType } from '@prisma/client';
-import { SellerProductPromotionService } from '@project-lc/nest-modules-seller';
 import { LiveShoppingService } from '@project-lc/nest-modules-liveshopping';
+import { ProductPromotionService } from '@project-lc/nest-modules-product-promotion';
 import { FirstmallDbService } from '../firstmall-db.service';
 import { StatCounter } from './utills/statCounter';
 @Injectable()
 export class FmOrdersService {
   constructor(
     private readonly db: FirstmallDbService,
-    private readonly sellerProductPromotionService: SellerProductPromotionService,
+    private readonly productPromotionService: ProductPromotionService,
     private readonly liveShoppingService: LiveShoppingService,
   ) {}
 
@@ -276,7 +276,7 @@ export class FmOrdersService {
 
     // * 판매유형(라이브쇼핑, 홍보페이지, 일반)
     const lvs = await this.liveShoppingService.findLiveShoppingsByGoodsIds(goodsIds);
-    const pps = await this.sellerProductPromotionService.findProductPromotionsByGoodsIds(
+    const pps = await this.productPromotionService.findProductPromotionsByGoodsIds(
       goodsIds,
     );
     const liveShoppingIds = lvs.map((x) => x.fmGoodsSeq);
