+++ libs/nest-modules-seller/src/index.ts
@@ -1,4 +1,3 @@
 export * from './seller/seller.module';
 export * from './seller/seller.service';
 export * from './seller/seller-settlement.service';
-export * from './seller/seller-product-promotion.service';
