+++ libs/nest-modules-seller/src/seller/seller.module.ts
@@ -1,11 +1,10 @@
-import { Module, DynamicModule } from '@nestjs/common';
+import { DynamicModule, Module } from '@nestjs/common';
+import { LiveShoppingModule } from '@project-lc/nest-modules-liveshopping';
 import { MailModule } from '@project-lc/nest-modules-mail';
 import { S3Module } from '@project-lc/nest-modules-s3';
-import { LiveShoppingModule } from '@project-lc/nest-modules-liveshopping';
 import { SellerContactsController } from './seller-contacts.controller';
 import { SellerContactsService } from './seller-contacts.service';
 import { SellerSettlementService } from './seller-settlement.service';
-import { SellerProductPromotionService } from './seller-product-promotion.service';
 import { SellerShopService } from './seller-shop.service';
 import { SellerController } from './seller.controller';
 import { SellerService } from './seller.service';
@@ -16,15 +15,10 @@ export class SellerModule {
     SellerService,
     SellerSettlementService,
     SellerShopService,
-    SellerProductPromotionService,
     SellerContactsService,
   ];
 
-  private static readonly exports = [
-    SellerService,
-    SellerSettlementService,
-    SellerProductPromotionService,
-  ];
+  private static readonly exports = [SellerService, SellerSettlementService];
 
   private static readonly controllers = [SellerController, SellerContactsController];
